Narrow Connector validation host adapters
