+++ src/tool_runtime/validation_profile/mod.rs
@@ -2,8 +2,10 @@
 mod recipe_tests;
 
 pub(crate) use webcodex_validation::{
-    resolve_validation_recipe, validation_adapter_for_tool, RecipeId, SemanticCheck,
-    ValidationAdapter, ValidationCommandOptions, ValidationFailureEvidence,
+    validation_adapter_for_tool, ValidationAdapter, ValidationCommandOptions,
 };
 #[cfg(test)]
-pub(crate) use webcodex_validation::{RecipeError, ResolvedValidationRecipe};
+pub(crate) use webcodex_validation::{
+    resolve_validation_recipe, RecipeError, RecipeId, ResolvedValidationRecipe, SemanticCheck,
+    ValidationFailureEvidence,
+};
